streamlined auth process
switched from express-basic auth package to basic-auth package. EBA acted more like a black box where there was no transparency. this meant that i i didnt have access to the request object when i was performing the actual validation. this lead to having to read the same record from the db again to add custom properties to the request object. BA is a single function, rather than a standalone middleware, which just decodes the username and password. this meant that i could manually perform the validation and add custom attributes to the request object, without having to read from the db twice